OCD-1410: Allow mass reject of pending surveillances

+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/impl/SurveillanceManagerImpl.java
@@ -20,6 +20,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+
 import gov.healthit.chpl.auth.SendMailUtil;
 import gov.healthit.chpl.auth.Util;
 import gov.healthit.chpl.auth.dao.UserPermissionDAO;
@@ -29,9 +31,12 @@ import gov.healthit.chpl.auth.permission.UserPermissionRetrievalException;
 import gov.healthit.chpl.caching.CacheNames;
 import gov.healthit.chpl.caching.ClearBasicSearch;
 import gov.healthit.chpl.dao.CertifiedProductDAO;
+import gov.healthit.chpl.dao.ContactDAO;
+import gov.healthit.chpl.dao.EntityCreationException;
 import gov.healthit.chpl.dao.EntityRetrievalException;
 import gov.healthit.chpl.dao.SurveillanceDAO;
 import gov.healthit.chpl.domain.CertifiedProduct;
+import gov.healthit.chpl.domain.Contact;
 import gov.healthit.chpl.domain.Surveillance;
 import gov.healthit.chpl.domain.SurveillanceNonconformity;
 import gov.healthit.chpl.domain.SurveillanceNonconformityDocument;
@@ -40,8 +45,10 @@ import gov.healthit.chpl.domain.SurveillanceRequirement;
 import gov.healthit.chpl.domain.SurveillanceRequirementType;
 import gov.healthit.chpl.domain.SurveillanceResultType;
 import gov.healthit.chpl.domain.SurveillanceType;
+import gov.healthit.chpl.domain.concept.ActivityConcept;
 import gov.healthit.chpl.dto.CertificationBodyDTO;
 import gov.healthit.chpl.dto.CertifiedProductDetailsDTO;
+import gov.healthit.chpl.dto.ContactDTO;
 import gov.healthit.chpl.entity.PendingSurveillanceEntity;
 import gov.healthit.chpl.entity.PendingSurveillanceNonconformityEntity;
 import gov.healthit.chpl.entity.PendingSurveillanceRequirementEntity;
@@ -50,8 +57,10 @@ import gov.healthit.chpl.entity.SurveillanceNonconformityDocumentationEntity;
 import gov.healthit.chpl.entity.SurveillanceNonconformityEntity;
 import gov.healthit.chpl.entity.SurveillanceRequirementEntity;
 import gov.healthit.chpl.entity.listing.CertifiedProductEntity;
+import gov.healthit.chpl.manager.ActivityManager;
 import gov.healthit.chpl.manager.SurveillanceManager;
 import gov.healthit.chpl.validation.surveillance.SurveillanceValidator;
+import gov.healthit.chpl.web.controller.exception.ObjectMissingValidationException;
 
 @Service
 public class SurveillanceManagerImpl implements SurveillanceManager {
@@ -59,10 +68,13 @@ public class SurveillanceManagerImpl implements SurveillanceManager {
 	@Autowired private SendMailUtil sendMailService;
 	@Autowired private Environment env;
 	
+	@Autowired ContactDAO contactDao;
 	@Autowired SurveillanceDAO survDao;
 	@Autowired CertifiedProductDAO cpDao;
 	@Autowired SurveillanceValidator validator;
 	@Autowired UserPermissionDAO userPermissionDao;
+	
+	@Autowired private ActivityManager activityManager;
 	
 	@Override
 	@Transactional(readOnly = true)
@@ -242,7 +254,7 @@ public class SurveillanceManagerImpl implements SurveillanceManager {
 	@PreAuthorize("(hasRole('ROLE_ACB_STAFF') or hasRole('ROLE_ACB_ADMIN')) "
 			+ "and hasPermission(#acbId, 'gov.healthit.chpl.dto.CertificationBodyDTO', admin)")
 	public Surveillance getPendingById(Long acbId, Long survId) throws EntityNotFoundException {
-		PendingSurveillanceEntity pending = survDao.getPendingSurveillanceById(survId);
+		PendingSurveillanceEntity pending = survDao.getPendingSurveillanceById(survId, false);
 		if(pending == null) {
 			throw new EntityNotFoundException("Could not find pending surveillance with id " + survId);
 		}
@@ -271,23 +283,46 @@ public class SurveillanceManagerImpl implements SurveillanceManager {
 	@Transactional
 	@PreAuthorize("(hasRole('ROLE_ACB_STAFF') or hasRole('ROLE_ACB_ADMIN')) "
 			+ "and hasPermission(#acbId, 'gov.healthit.chpl.dto.CertificationBodyDTO', admin)")
-	public void deletePendingSurveillance(Long acbId, Long survId) {		
-		Surveillance surv = new Surveillance();
-		surv.setId(survId);
+	public void deletePendingSurveillance(Long acbId, Long survId, boolean isConfirmed) throws ObjectMissingValidationException, JsonProcessingException, EntityRetrievalException, EntityCreationException {		
+		PendingSurveillanceEntity surv = survDao.getPendingSurveillanceById(survId, true);
+		if(surv == null) {
+			throw new EntityNotFoundException("Could not find pending surveillance with id " + survId);
+		}
+		CertifiedProductEntity ownerCp = surv.getCertifiedProduct();
+		if(ownerCp == null) {
+			throw new EntityNotFoundException("Could not find certified product associated with pending surveillance.");
+		}
+		
+		Surveillance toDelete = getPendingById(acbId, survId);
 		
 		try {
-			survDao.deletePendingSurveillance(surv);
+			survDao.deletePendingSurveillance(toDelete);
 		} catch(Exception ex) {
 			logger.error("Error marking pending surveillance with id " + survId + " as deleted.", ex);
+		}
+		
+		if(isPendingSurveillanceAvailableForUpdate(ownerCp.getCertificationBodyId(), surv)) {
+			try {
+				survDao.deletePendingSurveillance(toDelete);
+			} catch(Exception ex) {
+				logger.error("Error marking pending surveillance with id " + toDelete.getId() + " as deleted.", ex);
+			}
+			StringBuilder activityMsg = new StringBuilder().append("Pending surveillance " + toDelete.getId() + " has been ");
+			if(isConfirmed){
+				activityMsg.append("confirmed.");
+			} else{
+				activityMsg.append("rejected.");
+			}
+			activityManager.addActivity(ActivityConcept.ACTIVITY_CONCEPT_PENDING_SURVEILLANCE, toDelete.getId(), activityMsg.toString(), toDelete, null);	
 		}
 	}
 	
 	@Override
 	@Transactional
 	@PreAuthorize("hasRole('ROLE_ACB_STAFF') or hasRole('ROLE_ACB_ADMIN')")
-	public void deletePendingSurveillance(List<CertificationBodyDTO> userAcbs, Long survId)
-		throws EntityNotFoundException, AccessDeniedException {
-		PendingSurveillanceEntity surv = survDao.getPendingSurveillanceById(survId);
+	public void deletePendingSurveillance(List<CertificationBodyDTO> userAcbs, Long survId, boolean isConfirmed)
+		throws EntityNotFoundException, AccessDeniedException, ObjectMissingValidationException, JsonProcessingException, EntityRetrievalException, EntityCreationException {
+		PendingSurveillanceEntity surv = survDao.getPendingSurveillanceById(survId, true);
 		if(surv == null) {
 			throw new EntityNotFoundException("Could not find pending surveillance with id " + survId);
 		}
@@ -308,14 +343,41 @@ public class SurveillanceManagerImpl implements SurveillanceManager {
 			throw new AccessDeniedException("Permission denied on ACB " + ownerCp.getCertificationBodyId() + " for user " + Util.getCurrentUser().getSubjectName());
 		}
 		
-		Surveillance toDelete = new Surveillance();
-		toDelete.setId(survId);
-		
-		try {
-			survDao.deletePendingSurveillance(toDelete);
-		} catch(Exception ex) {
-			logger.error("Error marking pending surveillance with id " + survId + " as deleted.", ex);
+		if(isPendingSurveillanceAvailableForUpdate(ownerCp.getCertificationBodyId(), surv)) {
+			Surveillance toDelete = getPendingById(ownerCp.getCertificationBodyId(), survId);
+			try {
+				survDao.deletePendingSurveillance(toDelete);
+			} catch(Exception ex) {
+				logger.error("Error marking pending surveillance with id " + toDelete.getId() + " as deleted.", ex);
+			}
+			
+			StringBuilder activityMsg = new StringBuilder().append("Pending surveillance " + toDelete.getId() + " has been ");
+			if(isConfirmed){
+				activityMsg.append("confirmed.");
+			} else{
+				activityMsg.append("rejected.");
+			}
+			activityManager.addActivity(ActivityConcept.ACTIVITY_CONCEPT_PENDING_SURVEILLANCE, toDelete.getId(), activityMsg.toString(), toDelete, null);
 		}
+	}
+	
+	@Override
+	@Transactional
+	@PreAuthorize("(hasRole('ROLE_ACB_STAFF') or hasRole('ROLE_ACB_ADMIN')) "
+			+ "and hasPermission(#acbId, 'gov.healthit.chpl.dto.CertificationBodyDTO', admin)")
+	public boolean isPendingSurveillanceAvailableForUpdate(Long acbId, PendingSurveillanceEntity pendingSurv) 
+	throws EntityRetrievalException, ObjectMissingValidationException {
+		if(pendingSurv.getDeleted().booleanValue() == true) {
+			ContactDTO contactDTO = contactDao.getById(pendingSurv.getLastModifiedUser());
+			ObjectMissingValidationException alreadyDeletedEx = new ObjectMissingValidationException();
+			alreadyDeletedEx.getErrorMessages().add("This pending surveillance has already been confirmed or rejected by another user.");
+			alreadyDeletedEx.setContact(contactDTO != null ? new Contact(contactDTO) : null);
+			alreadyDeletedEx.setObjectId(pendingSurv.getId().toString());
+			alreadyDeletedEx.setStartDate(pendingSurv.getStartDate());
+			alreadyDeletedEx.setEndDate(pendingSurv.getEndDate());
+			throw alreadyDeletedEx;
+		}
+		return pendingSurv != null;
 	}
 	
 	@Override
